requirements.txt
certifi==2020.12.5
chardet==4.0.0
falcon==2.0.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
marshmallow==3.9.1
psycopg2-binary==2.8.6
requests==2.25.1
urllib3==1.26.2
